+++ lib/Pages/Main_Screens/home.dart
@@ -1,11 +1,7 @@
+import 'package:biscuit2/Post2.dart';
 import 'package:biscuit2/post1.dart';
 import 'package:biscuit2/post3.dart';
 import 'package:flutter/material.dart';
-
-
-
-
-import 'Post2.dart';
 
 class UserHomePage extends StatelessWidget {
   final _controller = PageController();
